pyproject.toml
[project]
name = "backoffice_tramites"
version = "0.1.0"
description = "Backoffice de tramites - Microservice for San Felipe Government"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django>=6.0.2",
    "django-environ>=0.13.0",
    "psycopg2-binary>=2.9.11",
    "pillow>=11.0.0",
    "requests>=2.32.0",
]

[dependency-groups]
dev = [
    "django-debug-toolbar>=6.2.0",
    "factory-boy>=3.3.0",
    "pyright>=1.1.408",
    "pytest>=9.0.2",
    "pytest-django>=4.12.0",
    "pytest-factoryboy>=2.8.1",
    "pytest-cov>=6.0.0",
    "ruff>=0.15.2",
]
prod = [
    "gunicorn>=25.1.0",
]

# =============================================================================
# Ruff Configuration
# =============================================================================
[tool.ruff]
target-version = "py314"
line-length = 100

[tool.ruff.lint]
# Enable these rule sets for Django projects
select = [
    "E",     # pycodestyle errors
    "W",     # pycodestyle warnings
    "F",     # pyflakes
    "I",     # isort
    "B",     # flake8-bugbear
    "C4",    # flake8-comprehensions
    "UP",    # pyupgrade
    "ASYNC", # flake8-async
    "PIE",   # flake8-pie
    "T20",   # flake8-print
    "RSE",   # flake8-raise
    "RET",   # flake8-return
    "SIM",   # flake8-simplify
    "PTH",   #7 flake8-use-pathlib
    "PGH",   # pygrep-hooks
    "PL",    # pylint
    "PERF",  # perflint
    "RUF",   # ruff-specific rules
]

ignore = [
    # Django often uses lowercase models
    "PLR2004",  # Magic value comparison
    # Allow _ in __init__.py for imports
    "RUF100",   # Unused noqa directive
    # Allow __init__.py to be empty sometimes
    "PLR2004",
    # Complex comprehensions can be readable
    "C416",
]

[tool.ruff.lint.per-file-ignores]
# Django settings files often have long lines
"sanfelipe/settings.py" = ["E501"]
# Tests can have long lines
"tests/**/*.py" = ["E501", "B011"]
# Allow _ in migrations
"migrations/**/*.py" = ["B008", "C4"]
# Allow unused imports in __init__.py
"__init__.py" = ["F401", "UP035"]

[tool.ruff.lint.isort]
# Known-first-party for Django apps
known-first-party = ["tramites", "catalogos", "bitacora", "costos", "core", "sanfelipe"]

[tool.ruff.lint.pydocstyle]
# Google-style docstrings (common in Django)
convention = "google"

[tool.ruff.format]
# Use single quotes for Python
quote-style = "single"

# =============================================================================
# Pytest Configuration
# =============================================================================
[tool.pytest.ini_options]
# Use test settings
DJANGO_SETTINGS_MODULE = 'sanfelipe.settings_test'
django_find_project = false

# Test discovery patterns
pythonpath = ['.']
testpaths = ['tests']
python_files = ['test_*.py', '*_test.py']
python_classes = ['Test*']
python_functions = ['test_*']

# Add default options
addopts = [
    '--reuse-db',
    '--tb=short',
    '--verbose',
]

# Pytest markers
markers = [
    'slow: marks tests as slow (deselect with -m "not slow")',
    'integration: marks tests as integration tests',
]
